move host into internal packages
